Refactors auto-completion
