+++ packages/language-support/src/autocompletion/helpers.ts
@@ -1,0 +1,147 @@
+import { ParserRuleContext } from 'antlr4';
+import { CodeCompletionCore } from 'antlr4-c3';
+import {
+  CompletionItem,
+  CompletionItemKind,
+} from 'vscode-languageserver-types';
+import { DbInfo } from '../dbInfo';
+import CypherParser, {
+  Expression2Context,
+  LabelExpression4Context,
+  LabelExpression4IsContext,
+  NodePatternContext,
+  ProcedureNameContext,
+  RelationshipPatternContext,
+} from '../generated-parser/CypherParser';
+import { findParent, isDefined } from '../helpers';
+import { CypherTokenType, lexerSymbols, tokenNames } from '../lexerSymbols';
+import { ParsingResult } from '../parserWrapper';
+
+export function isLabel(p: ParserRuleContext) {
+  return (
+    p instanceof LabelExpression4Context ||
+    p instanceof LabelExpression4IsContext
+  );
+}
+
+export function inNodeLabel(stopNode: ParserRuleContext) {
+  const labelParent = findParent(
+    findParent(stopNode, isLabel),
+    (p) => p instanceof NodePatternContext,
+  );
+
+  return isDefined(labelParent);
+}
+
+export function inRelationshipType(stopNode: ParserRuleContext) {
+  const relTypeParent = findParent(
+    findParent(stopNode, isLabel),
+    (p) => p instanceof RelationshipPatternContext,
+  );
+
+  return isDefined(relTypeParent);
+}
+
+export function parentExpression(stopNode: ParserRuleContext) {
+  return findParent(stopNode, (p) => p instanceof Expression2Context);
+}
+
+export function inProcedureName(stopNode: ParserRuleContext) {
+  const procParent = findParent(
+    stopNode,
+    (p) => p instanceof ProcedureNameContext,
+  );
+
+  return isDefined(procParent);
+}
+
+export function autocompleteLabels(dbInfo: DbInfo) {
+  return dbInfo.labels.map((t) => {
+    return {
+      label: t,
+      kind: CompletionItemKind.TypeParameter,
+    };
+  });
+}
+
+export function autocompleteRelTypes(dbInfo: DbInfo) {
+  return dbInfo.relationshipTypes.map((t) => {
+    return {
+      label: t,
+      kind: CompletionItemKind.TypeParameter,
+    };
+  });
+}
+
+export function autoCompleteFunctions(dbInfo: DbInfo, expr: ParserRuleContext) {
+  return Array.from(dbInfo.functionSignatures.keys())
+    .filter((functionName) => {
+      return functionName.startsWith(expr.getText());
+    })
+    .map((t) => {
+      return {
+        label: t,
+        kind: CompletionItemKind.Function,
+      };
+    });
+}
+
+export function autoCompleteProcNames(dbInfo: DbInfo) {
+  return Array.from(dbInfo.procedureSignatures.keys()).map((t) => {
+    return {
+      label: t,
+      kind: CompletionItemKind.Function,
+    };
+  });
+}
+
+export function autoCompleteKeywords(parsingResult: ParsingResult) {
+  const parser = parsingResult.parser;
+  const tokens = parsingResult.tokens;
+
+  const codeCompletion = new CodeCompletionCore(parser);
+
+  // We always need to subtract one more for the final EOF
+  const caretIndex = tokens.length - 2;
+
+  if (caretIndex >= 0) {
+    // We need this to ignore the list of tokens from:
+    // * unescapedSymbolicNameString, because a lot of keywords are allowed there
+    // * escapedSymbolicNameString, to avoid showing ESCAPED_SYMBOLIC_NAME
+    //
+    // That way we do not populate tokens that are coming from those rules and those
+    // are collected as rule names instead
+    codeCompletion.preferredRules = new Set<number>()
+      .add(CypherParser.RULE_unescapedSymbolicNameString)
+      .add(CypherParser.RULE_escapedSymbolicNameString);
+
+    // Keep only keywords as suggestions
+    codeCompletion.ignoredTokens = new Set<number>(
+      Object.entries(lexerSymbols)
+        .filter(([, type]) => type !== CypherTokenType.keyword)
+        .map(([token]) => Number(token)),
+    );
+
+    const candidates = codeCompletion.collectCandidates(caretIndex);
+    const tokens = candidates.tokens.entries();
+
+    const tokenCandidates = Array.from(tokens).map((value) => {
+      const [tokenNumber, followUpList] = value;
+      return [tokenNumber]
+        .concat(followUpList)
+        .map((value) => tokenNames[value])
+        .join(' ');
+    });
+
+    const tokenCompletions: CompletionItem[] = tokenCandidates.map((t) => {
+      return {
+        label: t,
+        kind: CompletionItemKind.Keyword,
+      };
+    });
+
+    return tokenCompletions;
+  } else {
+    return [];
+  }
+}
