+++ packages/language-support/src/helpers.ts
@@ -63,6 +63,10 @@ export function parse(cypher: string) {
   return parser.statements();
 }
 
+export function isDefined(x: unknown) {
+  return x !== null && x !== undefined;
+}
+
 type AntlrDefaultExport = {
   tree: {
     Trees: {
