+++ packages/language-support/src/index.ts
@@ -1,5 +1,5 @@
 export type { ParserRuleContext } from 'antlr4';
-export { autocomplete } from './autocompletion';
+export { autocomplete } from './autocompletion/autocompletion';
 export type { DbInfo } from './dbInfo';
 export { antlrUtils, parse } from './helpers';
 export {
